requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via httpx
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
bcrypt==4.2.0
    # via paramiko
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via typer
cosl==0.0.36
    # via -r requirements.in
cryptography==43.0.1
    # via
    #   -r requirements.in
    #   paramiko
durationpy==0.7
    # via kubernetes
google-auth==2.35.0
    # via kubernetes
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via lightkube
hvac==2.3.0
    # via juju
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via -r requirements.in
jsonschema==4.23.0
    # via -r requirements.in
jsonschema-specifications==2023.12.1
    # via jsonschema
juju==3.5.2.0
    # via -r requirements.in
kubernetes==31.0.0
    # via juju
lightkube==0.15.4
    # via cosl
lightkube-models==1.31.1.8
    # via lightkube
macaroonbakery==1.3.4
    # via juju
markupsafe==2.1.5
    # via jinja2
mypy-extensions==1.0.0
    # via typing-inspect
oauthlib==3.2.2
    # via
    #   kubernetes
    #   requests-oauthlib
ops==2.17.0
    # via
    #   -r requirements.in
    #   cosl
    #   ops-scenario
ops-scenario==7.0.5
    # via pytest-interface-tester
packaging==24.1
    # via
    #   juju
    #   pytest
paramiko==3.5.0
    # via juju
pluggy==1.5.0
    # via pytest
protobuf==5.28.2
    # via macaroonbakery
pyasn1==0.6.1
    # via
    #   juju
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   cosl
    #   pytest-interface-tester
pydantic-core==2.23.4
    # via pydantic
pymacaroons==0.13.0
    # via macaroonbakery
pynacl==1.5.0
    # via
    #   macaroonbakery
    #   paramiko
    #   pymacaroons
pyrfc3339==1.1
    # via
    #   juju
    #   macaroonbakery
pytest==8.3.3
    # via pytest-interface-tester
pytest-interface-tester==3.2.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via kubernetes
pytz==2024.2
    # via pyrfc3339
pyyaml==6.0.2
    # via
    #   cosl
    #   juju
    #   kubernetes
    #   lightkube
    #   ops
    #   ops-scenario
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   hvac
    #   kubernetes
    #   macaroonbakery
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via kubernetes
rpds-py==0.18.1
    # via
    #   -r requirements.in
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   kubernetes
    #   macaroonbakery
    #   pymacaroons
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
tenacity==9.0.0
    # via cosl
toposort==1.10
    # via juju
typer==0.7.0
    # via pytest-interface-tester
typing-extensions==4.12.2
    # via
    #   cosl
    #   pydantic
    #   pydantic-core
    #   typing-inspect
typing-inspect==0.9.0
    # via juju
urllib3==2.2.3
    # via
    #   kubernetes
    #   requests
websocket-client==1.8.0
    # via
    #   kubernetes
    #   ops
websockets==13.1
    # via juju
